+++ packages/nats-jetstream/src/index.ts
@@ -3,6 +3,7 @@ export * from './lib/jetstream.transport';
 export * from './lib/nats.client';
 export * from './lib/nats.context';
 export * from './lib/nats.constants';
+export * from './lib/jetstream.types';
 export * from './lib/interfaces/nats-client-options.interface';
 export * from './lib/interfaces/nats-jetstream-options.interface';
 export * from './lib/interfaces/nats-stream-config.interface';